refactor code to extract logic into constants
