+++ packages/format-cnpj/src/constants.js
@@ -1,0 +1,7 @@
+export const CNPJ_LENGTH = 14;
+
+export const DOT_INDEXES = [1, 4];
+
+export const SLASH_INDEXES = [7];
+
+export const HYPHEN_INDEXES = [11];
