+++ packages/format-cnpj/src/index.js
@@ -1,21 +1,24 @@
 import onlyNumbers from '@brazilian-utils/helper-only-numbers';
 
-export const CNPJ_LENGTH = 14;
+import { CNPJ_LENGTH, DOT_INDEXES, SLASH_INDEXES, HYPHEN_INDEXES } from './constants';
+
+const isLastDigit = (cnpj, index) => index === cnpj.length - 1;
 
 export default function formatCnpj(cnpj) {
   if (!cnpj) return '';
 
   const numericCNPJ = onlyNumbers(cnpj);
 
-  return numericCNPJ
-    .split('')
-    .map((digit, i) => {
-      if (i !== numericCNPJ.length - 1) {
-        if (i === 1 || i === 4) return `${digit}.`;
-        if (i === 7) return `${digit}/`;
-        if (i === 11) return `${digit}-`;
+  const arrayCNPJ = numericCNPJ.split('');
+
+  return arrayCNPJ
+    .map((digit, index) => {
+      if (!isLastDigit(numericCNPJ, index)) {
+        if (DOT_INDEXES.includes(index)) return `${digit}.`;
+        if (SLASH_INDEXES.includes(index)) return `${digit}/`;
+        if (HYPHEN_INDEXES.includes(index)) return `${digit}-`;
       }
-      return i < CNPJ_LENGTH ? digit : '';
+      return index < CNPJ_LENGTH ? digit : '';
     })
     .join('');
 }
