+++ packages/format-cnpj/src/tests/index.test.js
@@ -1,4 +1,5 @@
-import formatCnpj, { CNPJ_LENGTH } from '../';
+import formatCnpj from '../';
+import { CNPJ_LENGTH } from '../constants';
 
 describe('formatCnpj', () => {
   test('should format cnpj with mask', () => {
