+++ packages/format-cpf/src/constants.js
@@ -1,0 +1,5 @@
+export const CPF_LENGTH = 11;
+
+export const DOT_INDEXES = [2, 5];
+
+export const HYPHEN_INDEXES = [8];
