+++ packages/format-cpf/src/index.js
@@ -1,20 +1,23 @@
 import onlyNumbers from '@brazilian-utils/helper-only-numbers';
 
-export const CPF_LENGTH = 11;
+import { CPF_LENGTH, DOT_INDEXES, HYPHEN_INDEXES } from './constants';
+
+const isLastDigit = (cpf, index) => index === cpf.length - 1;
 
 export default function formatCpf(cpf) {
   if (!cpf) return '';
 
   const numericCPF = onlyNumbers(cpf);
 
-  return numericCPF
-    .split('')
-    .map((digit, i) => {
-      if (i !== numericCPF.length - 1) {
-        if (i === 2 || i === 5) return `${digit}.`;
-        if (i === 8) return `${digit}-`;
+  const arrayCPF = numericCPF.split('');
+
+  return arrayCPF
+    .map((digit, index) => {
+      if (!isLastDigit(numericCPF, index)) {
+        if (DOT_INDEXES.includes(index)) return `${digit}.`;
+        if (HYPHEN_INDEXES.includes(index)) return `${digit}-`;
       }
-      return i < CPF_LENGTH ? digit : '';
+      return index < CPF_LENGTH ? digit : '';
     })
     .join('');
 }
